[FEATURE] Video Session (#529)
* Introducing Video Upload and Video Player in Session

Renaming audio components to media components
Enabling mp4 for upload
Adding div container as test container for video player

* First Working Video Display Version

* Added Delete Media Option in UI and API

Delete operation is also successful if file does not exist any more, but media entity is still on DB.
With the deletion of the media entity, the session reference is set to null.
Renamed the session-audio component to session-media.

* Added NVM Use v18 in Husky Precommit

* First Version of URL Video Maintenance Enablement in Session Maintenance UI

Enhanced Media resolver with new create external media endpoint accepting the url.
Using external as mimeType and the existing "file" name to store the url
Added a cleanup in case a new video is uploaded and there was already a media connection in that session (otherwise there would be dangling media files which cannot be deleted manually by the user any more)
Added rudimentary UI control for URL input. Reacting with the other controls on the current state of each other (if url is filled, file upload is disabled, if file upload was used, the url input is disabled)

* Added Basic SwitchTube Support in Video Session

New SwitchTube WaveSurfer-like handling of interaction with iFrame embedding of video. External embedding works via URL only (video tag is created the same way the internal video upload works)

* Fixed Start/Pause Button with SwitchTube WaveSurfer

* Introduced Youtube Video Session Support

Added Wavesurfer version of Youtube embedded videos
Differentiated external mimeTypes in youtube and switchtube

* Switched to Delta Update for Media Id Update of Session

Fixed Deletion of Media in case of removing URL value

* Added Video Timeline

Centered Video player and increased relative size of the video container

* Hid Media Control for Video Sessions

In case of video sessions the control (play, pause, stop) should be done through the standard video controls.

* Fixed Centering the Video Player on the Session Page

* Enable Browser URL Media Instead of Embedding URL

Transforming from current youtube and switchtube browser address format into the needed iframe "embed" URL

* Fixed Unnecessary Media Download Request for External Video Sessions

* Added Participant Handling for Video Sessions

Removed video from participants session page in case there is no "local" participant usage switched on (owner led session)
Remembering the length of media in the session entity to still be able to display annotations after media was deleted

* Added Live Session Checkbox in Session Settings

Checking it does not yet have any effects, but it is stored in the database

* Added Logic to Hide Media Tab If Live Session Checkbox Is Selected

In case the live session checkbox is selected and there was already a media upload, the file is deleted and a deletion message is shown to the user.

* Switched to LiveSession Checks Instead of AudioSession Checks

* Added Support for Videos in Session Analysis Page

* Fixed Enablement of Marker Buttons in Participants Video Session

* Added Deletion of Media on Archiving

Fixed own video file play issue that wavesurfer is not used correctly.
Added picture-in-picture option to iframe option of youtube and switchtube

* Added Separation of Media Upload and URL Input

Enablement switches depending on the current state. If media file is uploaded, URL card/section is disabled and the other way round.

* Added Live Session UI Test

Fixed existing tests because of non-existing features (such as the save button in the session maintain page) or how the status of the session is selected

* Added Delete Media File Confirmation Dialog

Added simple confirmation dialog test in UI e2e tests

* Added Fix for Non-Existing Media Delete Handling

* Fixed Competing MailServer Port Issue By Respecting Config Mail Port

In certain scenarios port 1025 might be blocked for MailDev. In this case, the configuration setting for port in the test environment shall be respected.

* Added Marker and Participant Session Maintain UI Tests

---------

+++ apps/api/src/app/media/media.service.ts
@@ -6,6 +6,8 @@ import { CreateMediaInput } from './dto/create-media.input';
 import { Media } from './entities/media.entity';
 import { FileUpload } from 'graphql-upload';
 import { ensureMediaFolder, getFilePath } from './media.util';
+import { unlink } from 'node:fs/promises';
+import { CreateExternalMediaInput } from './dto/create-external-media.input';
 
 @Injectable()
 export class MediaService {
@@ -24,6 +26,28 @@ export class MediaService {
     const media = await this.mediaRepository.save(newMedia);
 
     await this.writeMediaFile(media.id, file);
+
+    return media;
+  }
+
+  async createExternal(createExternalMediaInput: CreateExternalMediaInput): Promise<Media> {
+    const newMedia = this.mediaRepository.create();
+    const url = createExternalMediaInput.url;
+
+    const youtubeMatch = url.match(/youtube\.com\/(?:watch\?(?:.*&)?v=|embed\/)([a-zA-Z0-9_-]+)/);
+    const switchtubeMatch = url.match(/tube\.switch\.ch\/(?:videos|embed)\/([a-zA-Z0-9_-]+)/);
+
+    if (youtubeMatch) {
+      newMedia.name = `https://www.youtube.com/embed/${youtubeMatch[1]}`;
+      newMedia.mimeType = 'external/youtube';
+    } else if (switchtubeMatch) {
+      newMedia.name = `https://tube.switch.ch/embed/${switchtubeMatch[1]}`;
+      newMedia.mimeType = 'external/switchtube';
+    } else {
+      throw new BadRequestException('Unsupported media type');
+    }
+
+    const media = await this.mediaRepository.save(newMedia);
 
     return media;
   }
@@ -54,7 +78,40 @@ export class MediaService {
     return this.mediaRepository.findOneByOrFail({ id });
   }
 
-  remove(id: number) {
-    return this.mediaRepository.delete(id);
+  async remove(id: number) {
+    const media = await this.findOne(id);
+
+    if (!media) {
+      throw new BadRequestException('Media not found');
+    }
+
+    await this.deleteMediaFile(id, media.name);
+
+    await this.mediaRepository.delete(id);
+
+    return media;
+  }
+
+  private async deleteMediaFile(id: number, filename: string): Promise<void> {
+    return new Promise(async (resolve, reject) => {
+      const filePath = getFilePath(id, filename);
+
+      try {
+        await unlink(filePath);
+
+        console.log(`Successfully deleted ${filePath}`);
+
+        resolve();
+      } catch (error: any) {
+        if (error.code === 'ENOENT') {
+          console.log(`File ${filePath} does not exist, skipping deletion.`);
+
+          resolve();
+        } else {
+          console.error(`Error deleting file ${filePath}:`, error);
+          reject(new BadRequestException('Could not delete file'));
+        }
+      }
+    });
   }
 }
